pyproject.toml
#
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "parensnet"
version = "0.1.0"
description = "Parallel Ensemble Network"
authors = [{ name = "Sriram P Chockalingam", email = "srirampc@gatech.edu" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[tool.poetry]
license = "Apache-2.0"
keywords = ["Data-driven", "Computational Biology", "Gene Regulatory Networks"]
packages = [{ include = "parensnet", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
devtools = "^0.12"
numpy = "^2.2"
pandas = "^2.3"
polars = "^1.31"
pydantic = "^2.11"
pyyaml = "^6.0"
scipy = "^1.15"
scikit-learn = "^1.7"
#
anndata = "^0.11"
arboreto = "^0.1"
lightgbm = "^4.6"
scanpy = "^1.11"
xgboost = "^3.0"

[tool.setuptools.packages.find]
include = ["parensnet"]

[tool.pyright]
include = ["parens", "eval", "data"]
exclude = ["**/__pycache__"]

[tool.black]
line-length = 88
target-version = ["py310"]
